+++ src/plonemeeting/portal/core/content/publication.py
@@ -9,24 +9,37 @@ from plone.app.contenttypes.content import File
 from plone.app.contenttypes.interfaces import IFile
 from plone.app.dexterity.textindexer import searchable
 from plone.app.textfield import RichText
+from plone.app.z3cform.widgets.contentbrowser import ContentBrowserFieldWidget
+from plone.app.z3cform.widgets.datetime import DatetimeFieldWidget
 from plone.app.z3cform.widgets.richtext import RichTextFieldWidget
+from plone.autoform.directives import no_omit
+from plone.autoform.directives import omitted
 from plone.autoform.directives import order_after
 from plone.autoform.directives import read_permission
 from plone.autoform.directives import widget
 from plone.autoform.directives import write_permission
 from plone.dexterity.content import Container
 from plone.indexer.decorator import indexer
+from plone.memoize.instance import memoize
 from plone.namedfile.field import NamedBlobFile
 from plone.supermodel import model
 from plonemeeting.portal.core import _
+from plonemeeting.portal.core.utils import get_linked_items_chain
+from plonemeeting.portal.core.utils import user_has_any_role
 from Products.CMFCore.permissions import ManagePortal
 from Products.CMFCore.permissions import ModifyPortalContent
+from Products.CMFCore.permissions import ReviewPortalContent
 from Products.CMFCore.permissions import View
 from Products.CMFCore.utils import _checkPermission
+from z3c.form.interfaces import IAddForm
+from z3c.form.interfaces import IEditForm
+from z3c.relationfield import RelationChoice
+from z3c.relationfield import RelationList
 from zope import schema
 from zope.component import getMultiAdapter
 from zope.globalrequest import getRequest
 from zope.interface import implementer
+from zope.schema import Datetime
 
 
 class IPublication(model.Schema, IFile, ITimestampableDocument):
@@ -72,7 +85,7 @@ class IPublication(model.Schema, IFile, ITimestampableDocument):
     model.primary("file")
     file = NamedBlobFile(title="File", accept=("application/pdf",), required=False)
 
-    # Styling fieldset
+    # Authority fieldset
     model.fieldset(
         "authority",
         label=_("Authority"),
@@ -99,6 +112,24 @@ class IPublication(model.Schema, IFile, ITimestampableDocument):
         description=_("expired_authority_date_description"),
         required=False,
     )
+
+    # Dates fieldset
+    # model.fieldset(
+    #     "dates",
+    #     label=_("Dates"),
+    #     fields=[
+    #         "is_archived_on_expiration_date"
+    #     ],
+    # )
+
+    # order_after(is_archived_on_expiration_date="IPublication.expires")
+    # # Let's put this after Plone's IPublication behavior's "expires" field
+    # is_archived_on_expiration_date = schema.Bool(
+    #     title=_("Archived on expiration date"),
+    #     description=_("is_archived_on_expiration_date_description"),
+    #     required=False,
+    #     default=False,
+    # )
 
     # Timestamping fieldset
     model.fieldset(
@@ -139,6 +170,8 @@ class Publication(Container, File):
     def is_timestamped(self):
         return ITimeStamper(self).is_timestamped()
 
+    # Workflow related methods
+
     def may_back_to_private(self):
         """Only Manager may back to private except if
         current review_state is "planned"."""
@@ -146,21 +179,39 @@ class Publication(Container, File):
         if api.content.get_state(self) == "planned":
             # Editor can't edit it directly, but can put it back to private to edit it.
             return user.has_role("Editor", object=self) or _checkPermission(ManagePortal, self)
+        if api.content.get_state(self) == "proposed":
+            # Reviewers can put it back to private.
+            return _checkPermission(ReviewPortalContent, self)
         else:
             return _checkPermission(ManagePortal, self)
 
     def may_plan(self):
-        """May plan if able to modify and
+        """May plan if able to review and
         a "publication date" (effectiveDate) is defined."""
-        return (
-            _checkPermission(ModifyPortalContent, self)
-            and self.effective_date is not None
-            and self.effective_date > DateTime()
-        )
+        if self.effective_date is None or self.effective_date <= DateTime():
+            # Can't plan if no effective date or effective date in the past
+            return False
+        return self.may_publish()  # Same guard as may_publish
 
     def may_publish(self):
         """May publish if able to modify."""
-        return _checkPermission(ModifyPortalContent, self)
+        state = api.content.get_state(self)
+        if user_has_any_role(["Manager", "Reviewer"], self):
+            # Managers and Reviewers can publish directly.
+            return True
+        if state == "private" and self._get_institution().has_publications_reviewers():
+            # Needs to be proposed first
+            return False
+        return _checkPermission(ReviewPortalContent, self)
+
+    def may_propose(self):
+        """May propose if able to modify and if validators are defined."""
+        institution = self._get_institution()
+        return institution.has_publications_reviewers() and _checkPermission(ReviewPortalContent, self)
+
+    def may_archive(self):
+        """May archive if manager."""
+        return user_has_any_role(["Manager"], self)
 
 
 @indexer(IPublication)