+++ src/plonemeeting/portal/core/events/meeting.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
+
 def meeting_state_changed(obj, event):
     items = obj.listFolderContents(contentFilter={"portal_type": "Item"})
     for item in items: